% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-LlmResponse-class.R
\name{as_table}
\alias{as_table}
\title{Generic extractor for LlmResponse outputs}
\usage{
as_table(x, ...)
}
\arguments{
\item{x}{An LlmResponse object}

\item{...}{Additional arguments (not used)}
}
\description{
Generic extractor for LlmResponse outputs
}
