% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-LlmResponse-class.R
\name{format.LlmResponse}
\alias{format.LlmResponse}
\title{Format an LlmResponse object}
\usage{
\method{format}{LlmResponse}(x, output_type = c("text", "meta", "logprobs", "complete"), ...)
}
\arguments{
\item{x}{An LlmResponse object}

\item{output_type}{A character string indicating the type of output to format.
Possible values are "text", "meta", "logprobs", or "complete".}

\item{...}{Additional arguments (not used)}
}
\value{
A formatted character string
}
\description{
Format an LlmResponse object
}
