% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-EllmerLlmApi-class.R
\name{get_llm_models.EllmerLlmApi}
\alias{get_llm_models.EllmerLlmApi}
\title{Retrieve Available LLM Models via Ellmer bridge}
\usage{
\method{get_llm_models}{EllmerLlmApi}(x, with_creds_only = TRUE, ...)
}
\arguments{
\item{x}{An EllmerLlmApi object}

\item{with_creds_only}{Logical, whether to attempt retrieval only if provider
supports it with credentials.}

\item{...}{Additional arguments}
}
\value{
Character vector or named list of available models
}
\description{
Retrieve Available LLM Models via Ellmer bridge
}
