% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-LocalLlmApi-class.R
\name{get_llm_models.LocalLlmApi}
\alias{get_llm_models.LocalLlmApi}
\title{Retrieve Available LLM Models}
\usage{
\method{get_llm_models}{LocalLlmApi}(x, ...)
}
\arguments{
\item{x}{An object of class LocalLlmApi}

\item{...}{Additional arguments}
}
\value{
A response object containing a list of available models from the selected API.
}
\description{
The get_llm_models() method fetches a list of available models from the local Ollama
Large Language Model (LLM) API.
It requires an LocalLlmApi object and returns the available models.
}
