% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-generics.R
\name{get_llm_models}
\alias{get_llm_models}
\title{Generic extractor for LlmApi outputs}
\usage{
get_llm_models(x, ...)
}
\arguments{
\item{x}{An LlmApi object}

\item{...}{Additional arguments}
}
\description{
Generic extractor for LlmApi outputs
}
