% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-LocalLlmApi-class.R
\name{get_llm_models_info.LocalLlmApi}
\alias{get_llm_models_info.LocalLlmApi}
\title{Retrieve Available LLM Models plus metadata}
\usage{
\method{get_llm_models_info}{LocalLlmApi}(x, ...)
}
\arguments{
\item{x}{An object of class LocalLlmApi}

\item{...}{Additional arguments}
}
\value{
A `LlmModelsInfo` object with `models` and selection metadata.
}
\description{
Retrieve Available LLM Models plus metadata
}
