% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-generics.R
\name{get_llm_models_info}
\alias{get_llm_models_info}
\title{Generic extractor for LLM model choices plus metadata}
\usage{
get_llm_models_info(x, ...)
}
\arguments{
\item{x}{An LlmApi object}

\item{...}{Additional arguments}
}
\description{
Generic extractor for LLM model choices plus metadata
}
