% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-RemoteLlmApi-class.R
\name{has_internet}
\alias{has_internet}
\title{Has Internet}
\usage{
has_internet(url = "https://www.google.com", timeout = 2)
}
\arguments{
\item{url}{(character) URL to test for internet connectivity. Default is "https://www.google.com".}

\item{timeout}{(numeric) number of seconds to wait for a response until giving up. Can not be less than 1 ms.}
}
\description{
Has Internet
}
