% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-LlmModelsInfo-class.R
\name{is_LlmModelsInfo}
\alias{is_LlmModelsInfo}
\title{Check if an object is of class LlmModelsInfo}
\usage{
is_LlmModelsInfo(x)
}
\arguments{
\item{x}{An object to check.}
}
\value{
Logical scalar indicating if the object is an LlmModelsInfo.
}
\description{
Check if an object is of class LlmModelsInfo
}
