% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-OllamaModelManager-helpers.R
\name{is_ollama_running}
\alias{is_ollama_running}
\title{Test Ollama Server Connection}
\usage{
is_ollama_running(url = Sys.getenv("OLLAMA_BASE_URL"))
}
\arguments{
\item{url}{Character string of the Ollama server URL.
Defaults to the value of the OLLAMA_BASE_URL environment variable.}
}
\value{
Logical TRUE if server is running, FALSE otherwise.
}
\description{
Checks if the local Ollama server is running and reachable.
}
