% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-LocalLlmApi-class.R
\name{new_LocalLlmApi}
\alias{new_LocalLlmApi}
\title{Create and Validate Local LLM API Credentials}
\usage{
new_LocalLlmApi(
  manager,
  new_model = "",
  base_url = Sys.getenv("OLLAMA_BASE_URL", unset = "http://localhost:11434"),
  excludePattern = ""
)
}
\arguments{
\item{manager}{An OllamaModelManager object}

\item{new_model}{Character, model name input from user (can be partial) of the model to pull}

\item{base_url}{Local Ollama base URL}

\item{excludePattern}{Character, a regex pattern to exclude certain models from the list of
available models, e.g. "babbage|curie|dall-e|davinci|text-embedding|tts|whisper"}
}
\value{
An object of class LocalLlmApi, or a list with an "error" attribute if construction fails.
}
\description{
Create and Validate Local LLM API Credentials
}
