% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-OllamaModelManager-class.R
\name{new_OllamaModelManager}
\alias{new_OllamaModelManager}
\title{Ollama Model Manager}
\usage{
new_OllamaModelManager()
}
\value{
An object of class OllamaModelManager
}
\description{
Manages the listing, checking, and pulling of Ollama models.
}
