% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-EllmerLlmApi-class.R
\name{print.EllmerLlmApi}
\alias{print.EllmerLlmApi}
\title{Print method for EllmerLlmApi}
\usage{
\method{print}{EllmerLlmApi}(x, ...)
}
\arguments{
\item{x}{An EllmerLlmApi object}

\item{...}{Additional arguments}
}
\description{
Print method for EllmerLlmApi
}
