% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-LlmPromptConfig-class.R
\name{print.LlmPromptConfig}
\alias{print.LlmPromptConfig}
\title{Print method for better readability}
\usage{
\method{print}{LlmPromptConfig}(x, ...)
}
\arguments{
\item{x}{An LlmPromptConfig object}

\item{...}{Additional arguments}
}
\description{
Print method for better readability
}
