% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-LlmPromptSettings-class.R
\name{print.LlmPromptSettings}
\alias{print.LlmPromptSettings}
\title{Print method for better readability}
\usage{
\method{print}{LlmPromptSettings}(x, ...)
}
\arguments{
\item{x}{An LlmPromptSettings object}

\item{...}{Additional arguments}
}
\description{
Print method for better readability
}
