% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-OllamaModel-class.R
\name{print.OllamaModel}
\alias{print.OllamaModel}
\title{Print method for OllamaModel}
\usage{
\method{print}{OllamaModel}(x, ...)
}
\arguments{
\item{x}{An OllamaModel object}

\item{...}{Additional arguments}
}
\description{
Print method for OllamaModel
}
