% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-LlmApi-class.R
\name{print.RemoteLlmApi}
\alias{print.RemoteLlmApi}
\title{Print method for LlmApi}
\usage{
\method{print}{RemoteLlmApi}(x, ...)
}
\arguments{
\item{x}{An LlmApi object}

\item{...}{Additional arguments}
}
\description{
Print method for LlmApi
}
