% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-EllmerLlmApi-class.R
\name{send_prompt.EllmerLlmApi}
\alias{send_prompt.EllmerLlmApi}
\title{Send a prompt through the Ellmer bridge}
\usage{
\method{send_prompt}{EllmerLlmApi}(api, prompt_config)
}
\arguments{
\item{api}{An EllmerLlmApi object}

\item{prompt_config}{An LlmPromptConfig object}
}
\value{
Normalized response with OpenAI-like `choices[[1]]$message$content`
}
\description{
Send a prompt through the Ellmer bridge
}
