% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-RemoteLlmApi-class.R
\name{send_prompt.RemoteLlmApi}
\alias{send_prompt.RemoteLlmApi}
\title{Send a prompt to a remote LLM API (e.g., OpenAI, DeepSeek)
This function sends a prompt to the remote LLM API and returns the response in a structured format.}
\usage{
\method{send_prompt}{RemoteLlmApi}(api, prompt_config)
}
\arguments{
\item{api}{An object of class RemoteLlmApi, which contains the API key and URL for the remote LLM API.}

\item{prompt_config}{An object of class LlmPromptConfig, containing the prompt content and model parameters.}
}
\value{
A list containing the response from the LLM API, structured similarly to OpenAI responses.
}
\description{
Send a prompt to a remote LLM API (e.g., OpenAI, DeepSeek)
This function sends a prompt to the remote LLM API and returns the response in a structured format.
}
\seealso{
[new_LlmResponse()]
}
