% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-OllamaModelManager-class.R
\name{update.OllamaModelManager}
\alias{update.OllamaModelManager}
\title{Update the list of locally available models}
\usage{
\method{update}{OllamaModelManager}(x, ...)
}
\arguments{
\item{x}{An OllamaModelManager object}

\item{...}{Not used, for compatibility}
}
\value{
Updated OllamaModelManager object
}
\description{
Update the list of locally available models
}
