% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-generics.R
\name{update}
\alias{update}
\title{Generic update function}
\usage{
update(x, ...)
}
\arguments{
\item{x}{Object to update}

\item{...}{Further arguments}
}
\description{
Dispatches update methods for different object classes.
}
