% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-LlmApi-class.R
\name{new_LocalLlmApi}
\alias{new_LocalLlmApi}
\title{Create a local LLM API object from model name and manager}
\usage{
new_LocalLlmApi(
  model_name,
  manager,
  base_url = "http://localhost:11434",
  pull_if_needed = TRUE
)
}
\arguments{
\item{model_name}{Character, model name input from user (can be partial)}

\item{manager}{An OllamaModelManager object}

\item{base_url}{Local Ollama base URL}

\item{pull_if_needed}{Logical, whether to pull the model automatically}
}
\value{
An object of class LocalLlmApi
}
\description{
Create a local LLM API object from model name and manager
}
