% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-generics.R
\name{send_prompt}
\alias{send_prompt}
\title{Generic LLM prompt sender
This function is a generic method for sending prompts to a remote or local LLM API.
It dispatches to the appropriate method based on the class of the `api` argument.}
\usage{
send_prompt(api, prompt_settings)
}
\arguments{
\item{api}{An object of class RemoteLlmApi or LocalLlmApi, which contains the API key and URL for the remote LLM API.}

\item{prompt_settings}{An object of class LlmPromptSettings, containing the prompt content and model parameters.}
}
\description{
Generic LLM prompt sender
This function is a generic method for sending prompts to a remote or local LLM API.
It dispatches to the appropriate method based on the class of the `api` argument.
}
